pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "researchclaw"
version = "1.0.0"
description = "ResearchClaw - chat-first research orchestrator with strict FSM and access policy"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6.0",
]

[project.optional-dependencies]
telegram = [
  "python-telegram-bot>=21.0",
]
dev = [
  "pytest>=8.0",
]

[tool.pytest.ini_options]
testpaths = ["researchclaw/tests"]
addopts = "-q"
